backend/requirements.txt
# AKA-Sim Backend Dependencies

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-socketio[asgi]>=5.9.0

# PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Data
numpy>=1.24.0
pydantic>=2.0.0

# Development
httpx>=0.24.0
